refactor: begin migrating profiler methods
